pyproject.toml
[project]
name = "dtu-work"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dspy>=2.6.24",
    "mlflow>=2.22.0",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
]
